+++ history.md
@@ -24,6 +24,7 @@
 # version x - t.b.d.
 - change '/api/info' to support readonly meta display
 - add /suggest/all to show all suggestions
+- upgrade dependencies to support Debian 10 (to fix: No usable version of the libssl was found)
 
 # version 0.1.5.5 - 2019-05-17
 - implement search suggestions API `/suggest?t=d`